requirements.txt
Django==4.2.7
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8
psycopg2-binary==2.9.9
djangorestframework==3.14.0